% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{broadcase.p.value}
\alias{broadcase.p.value}
\title{Sensitivity Analysis for Mantel-Haenszel Broad Case Test}
\usage{
broadcase.p.value(broad.case, matchvec, z, Gamma, two.sided = FALSE)
}
\arguments{
\item{broad.case}{A vector of broad case status with no missing data.}

\item{matchvec}{Matched set indicator, 1, 2, ..., sum(broad.case) with length(matchvec)==length(broad.case). Matched set indicators should be either integers or a factor.}

\item{z}{Treatment indicator, z=1 for treated, z=0 for control with length(z)==length(broad.case).}

\item{Gamma}{The sensitivity parameter \eqn{\Gamma} at which the test is conducted, where \eqn{\Gamma\ge 1}. Setting \eqn{\Gamma = 1} is equivalent to assuming exchangeability given the matched sets, and it performs a within-set randomization test.}

\item{two.sided}{One-sided or two-sided test. Default is FALSE.}
}
\value{
Upper bound on the p-value for all distributions of treatment assignment consistent with the sensitivity parameter Gamma.
}
\description{
Sensitivity Analysis for Mantel-Haenszel Broad Case Test
}
\details{
Sensitivity analysis for the Mantel-Haenszel broad case test \eqn{Y_b} of no treatment effect based on the broad case definition for matched case-control studies.
}
\examples{

dat<-data_gen(I=1e3, J=6, pi=1/3, bT=0.3, bC=0.1, etaT=0.3, etaC=0.15)
broadcase.p.value(dat$broad.case,dat$matchvec,dat$z,Gamma=3)

}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
