% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{cal_power}
\alias{cal_power}
\title{Calculate Power of Sensitivity Analysis (by formula)}
\usage{
cal_power(I, J, pi, bT, bC, etaT, etaC, Gamma, Theta, alpha)
}
\arguments{
\item{I}{Number of the broad case matched sets.}

\item{J}{Number of subjects in each matched sets.}

\item{pi}{Probability of receiving treatment.}

\item{bT}{Probability of being a broad case when treated.}

\item{bC}{Probability of being a broad case when not treated.}

\item{etaT}{Probability of a broad case being a narrow case when treated.}

\item{etaC}{Probability of a broad case being a narrow case when not treated.}

\item{Gamma}{The sensitivity parameter \eqn{\Gamma} at which the test is conducted, where \eqn{\Gamma\ge 1}. Setting \eqn{\Gamma = 1} is equivalent to assuming exchangeability given the matched sets, and it performs a within-set randomization test.}

\item{Theta}{The sensitivity parameter \eqn{\Theta} at which the test is conducted, where \eqn{\Theta\ge 1}. \eqn{1-\Theta} can be interpreted as the proportion of narrow cases when treated changing case definition when untreated. Setting \eqn{\Theta = 1} means that the treatment on average does not change case definition among always-cases.}

\item{alpha}{Significance level, usually 0.05.}
}
\value{
A matrix with calculated power results.
}
\description{
Calculate Power of Sensitivity Analysis (by formula)
}
\examples{

cal_power(I=1e3, J=6, pi=1/3, bT=0.3, bC=0.1, etaT=0.3, etaC=0.15, Gamma=3, Theta=1.5, alpha=0.05)

}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
