% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{cal_size_b}
\alias{cal_size_b}
\title{Calculate the Number of Broad Case Matched Sets to Achieve the Target Power Using the Broad Case Test}
\usage{
cal_size_b(pw, J, pi, bT, bC, etaT, etaC, Gamma, alpha, upper = 1e+07)
}
\arguments{
\item{pw}{Target power.}

\item{J}{Number of subjects in each matched sets.}

\item{pi}{Probability of receiving treatment.}

\item{bT}{Probability of being a broad case when treated.}

\item{bC}{Probability of being a broad case when not treated.}

\item{etaT}{Probability of a broad case being a narrow case when treated.}

\item{etaC}{Probability of a broad case being a narrow case when not treated.}

\item{Gamma}{The sensitivity parameter \eqn{\Gamma} at which the test is conducted, where \eqn{\Gamma\ge 1}. Setting \eqn{\Gamma = 1} is equivalent to assuming exchangeability given the matched sets, and it performs a within-set randomization test.}

\item{alpha}{Significance level, usually 0.05.}

\item{upper}{Upper bound on \eqn{I}. If the calculated power is below pw at the upper bound, output Inf.}
}
\value{
The number of broad case matched sets needed.
}
\description{
Calculate the Number of Broad Case Matched Sets to Achieve the Target Power Using the Broad Case Test
}
\examples{

cal_size_b(pw=0.8, J=6, pi=1/3, bT=0.3, bC=0.1, etaT=0.3, etaC=0.15, Gamma=3, alpha=0.05)

}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
