% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{data_gen}
\alias{data_gen}
\title{Generate simulation datasets.}
\usage{
data_gen(I, J, pi, bT, bC, etaT, etaC)
}
\arguments{
\item{I}{Number of the broad case matched sets.}

\item{J}{Number of subjects in each matched sets.}

\item{pi}{Probability of receiving treatment.}

\item{bT}{Probability of being a broad case when treated.}

\item{bC}{Probability of being a broad case when not treated.}

\item{etaT}{Probability of a broad case being a narrow case when treated.}

\item{etaC}{Probability of a broad case being a narrow case when not treated.}
}
\value{
A data frame.
}
\description{
Generate simulation datasets.
}
\examples{
data_gen(I=1e3, J=6, pi=1/3, bT=0.3, bC=0.1, etaT=0.3, etaC=0.15)

}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
