% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{design_sensitivity}
\alias{design_sensitivity}
\title{Design sensitivity formulas for the broad case and narrow case tests when data is generated as described in Section 4.2}
\usage{
design_sensitivity(J, pi, bT, bC, etaT, etaC)
}
\arguments{
\item{J}{Number of subjects in each matched sets.}

\item{pi}{Probability of receiving treatment.}

\item{bT}{Probability of being a broad case when treated.}

\item{bC}{Probability of being a broad case when not treated.}

\item{etaT}{Probability of a broad case being a narrow case when treated.}

\item{etaC}{Probability of a broad case being a narrow case when not treated.}
}
\value{
A matrix
\describe{
\item{broad.case}{The design sensitivity for the broad case test.}
\item{narrow.case}{The design sensitivity for the narrow case test.}
}
}
\description{
Design sensitivity formulas for the broad case and narrow case tests when data is generated as described in Section 4.2
}
\examples{
design_sensitivity(J,pi,bT,bC,etaT,etaC)


}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
