% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{narrowcase.p.value}
\alias{narrowcase.p.value}
\title{Sensitivity Analysis for Mantel-Haenszel Narrow Case Test}
\usage{
narrowcase.p.value(narrow.case, matchvec, z, Gamma, Theta,
  two.sided = FALSE)
}
\arguments{
\item{narrow.case}{A vector of narrow case status with no missing data.}

\item{matchvec}{Matched set indicator, 1, 2, ..., sum(narrow.case) with length(matchvec)==length(narrow.case). Matched set indicators should be either integers or a factor.}

\item{z}{Treatment indicator, z=1 for treated, z=0 for control with length(z)==length(narrow.case).}

\item{Gamma}{The sensitivity parameter \eqn{\Gamma} at which the test is conducted, where \eqn{\Gamma\ge 1}. Setting \eqn{\Gamma = 1} is equivalent to assuming exchangeability given the matched sets, and it performs a within-set randomization test.}

\item{Theta}{The sensitivity parameter \eqn{\Theta} at which the test is conducted, where \eqn{\Theta\ge 1}. \eqn{1-\Theta} can be interpreted as the proportion of narrow cases when treated changing case definition when untreated. Setting \eqn{\Theta = 1} means that the treatment on average does not change case definition among always-cases.}

\item{two.sided}{One-sided or two-sided test. Default is FALSE.}
}
\value{
Upper bound on the p-value for all distributions of treatment assignment consistent with the sensitivity parameter Gamma.
}
\description{
Sensitivity Analysis for Mantel-Haenszel Narrow Case Test
}
\details{
Sensitivity analysis for the Mantel-Haenszel broad case test \eqn{Y_b} of no treatment effect based on the broad case definition for matched case-control studies.
}
\examples{

dat<-data_gen(I=1e3, J=6, pi=1/3, bT=0.3, bC=0.1, etaT=0.3, etaC=0.15)
narrow.case.matched.index<-unique(subset(dat,narrow.case==1)$matchvec)
dat.narrow<-subset(dat,matchvec \%in\% narrow.case.matched.index)
narrowcase.p.value(dat.narrow$narrow.case,dat.narrow$matchvec,dat.narrow$z,Gamma=3,Theta=1.5)

}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
