% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_definition.R
\name{plot_power_byI}
\alias{plot_power_byI}
\title{Plot Power Against the Number of Broad Case Matched Sets}
\usage{
plot_power_byI(J, pi, bT, bC, etaT, etaC, alpha, n_sim = NULL,
  two.sided = FALSE, method = c("formula", "both"))
}
\arguments{
\item{J}{Number of subjects in each matched sets.}

\item{pi}{Probability of receiving treatment.}

\item{bT}{Probability of being a broad case when treated.}

\item{bC}{Probability of being a broad case when not treated.}

\item{etaT}{Probability of a broad case being a narrow case when treated.}

\item{etaC}{Probability of a broad case being a narrow case when not treated.}

\item{alpha}{Significance level, usually 0.05.}

\item{n_sim}{Number of simulation repetition.}

\item{two.sided}{One-sided or two-sided test. Default is FALSE.}

\item{method}{Whether to only include the formula power or both the formula and simulation power.}
}
\value{
A power plot saved at the working directory.
}
\description{
Plot Power Against the Number of Broad Case Matched Sets
}
\examples{

plot_power_byI(J=6, pi=1/3, bT=0.3, bC=0.1, etaT=0.3, etaC=0.15, alpha=0.05, method="formula")

}
\references{
Ting Ye and Dylan S. Small (2021). Combining Broad and Narrow Case Definitions in Matched Case-Control Studies.
}
